backend/requirements.txt
flask 
pymongo 
requests
bcrypt
python-dotenv